requirements.txt
streamlit==1.47.1
anthropic==0.61.0
python-dotenv==1.1.1 